+++ activerecord/lib/active_record/connection_adapters/sqlite3/quoting.rb
@@ -53,6 +53,10 @@ module ActiveRecord
           COLUMN_NAME_WITH_ORDER
         end
 
+        def column_name_without_alias_matcher
+          COLUMN_NAME_WITHOUT_ALIAS
+        end
+
         COLUMN_NAME = /
           \A
           (
@@ -79,7 +83,19 @@ module ActiveRecord
           \z
         /ix
 
-        private_constant :COLUMN_NAME, :COLUMN_NAME_WITH_ORDER
+        COLUMN_NAME_WITHOUT_ALIAS = /
+          \A
+          (
+            (?:
+              # "table_name"."column_name" | function(one or no argument)
+              ((?:\w+\.|"\w+"\.)?(?:\w+|"\w+")) | \w+\((?:|\g<2>)\)
+            )
+          )
+          (?:\s*,\s*\g<1>)*
+          \z
+        /ix
+
+        private_constant :COLUMN_NAME, :COLUMN_NAME_WITH_ORDER, :COLUMN_NAME_WITHOUT_ALIAS
 
         private
           def _type_cast(value)